split handler from phase
